+++ ui/packages/app/src/components/TabItem/Tabs.svelte
@@ -4,9 +4,8 @@
 	export let elem_id: string = "";
 	export let label: string;
 	export let id: string | number;
-	export let tab_index: number;
 </script>
 
-<TabItem {elem_id} name={label} {id} {tab_index} on:select>
+<TabItem {elem_id} name={label} {id} on:select>
 	<slot />
 </TabItem>